+++ src/lib/types/availability/update-params.type.ts
@@ -1,4 +1,6 @@
 export interface UpdateParams {
     reservations: number;
     available: number;
+    date: string;
+    time: string;
 }